extracted DefinitionListItem
