+++ components/vault/VaultInformation.tsx
@@ -1,10 +1,10 @@
 import { BigNumber } from 'bignumber.js'
 import { useTranslation } from 'next-i18next'
 import React from 'react'
-import { Card, Flex, Heading, Text } from 'theme-ui'
+import { Card, Heading, Text } from 'theme-ui'
 
+import { DefinitionList, DefinitionListItem } from '../DefinitionList'
 import { AppLink } from '../Links'
-import { ListDefinition } from '../ListDefinition'
 import { WithArrow } from '../WithArrow'
 
 interface VaultInformationItemProps {
@@ -14,21 +14,10 @@ interface VaultInformationItemProps {
 
 function VaultInformationItem({ text, value }: VaultInformationItemProps) {
   return (
-    <Flex
-      as="li"
-      sx={{
-        justifyContent: 'space-between',
-        py: 3,
-        borderBottom: '1px solid',
-        borderBottomColor: 'border',
-        '&:last-child': {
-          borderBottom: 'none',
-        },
-      }}
-    >
-      <Text sx={{ color: 'text.subtitle', fontWeight: 'semiBold', fontSize: 1 }}>{text}</Text>
-      <Text sx={{ color: 'primary', fontWeight: 'semiBold', fontSize: 1 }}>{value}</Text>
-    </Flex>
+    <DefinitionListItem sx={{ display: 'flex', justifyContent: 'space-between' }}>
+      <Text sx={{ color: 'text.subtitle' }}>{text}</Text>
+      <Text sx={{ color: 'primary' }}>{value}</Text>
+    </DefinitionListItem>
   )
 }
 
@@ -50,12 +39,11 @@ export function VaultInformation({ items }: VaultInformationProps) {
       <Heading variant="headerSettings" sx={{ mb: 3 }}>
         {t('vault-information')}
       </Heading>
-      <ListDefinition>
+      <DefinitionList>
         {items.map((item) => (
           <VaultInformationItem key={item.text} text={item.text} value={item.value} />
         ))}
-      </ListDefinition>
-
+      </DefinitionList>
       <AppLink href="https://kb.oasis.app/help">
         <WithArrow sx={{ color: 'link', fontSize: 1, mt: 3 }}>{t('learn-more-at-oasis')}</WithArrow>
       </AppLink>
