+++ features/vaultHistory/VaultHistoryView.tsx
@@ -15,7 +15,7 @@ import { useTranslation } from 'next-i18next'
 import React, { useState } from 'react'
 import { Box, Card, Flex, Grid, Heading, Text } from 'theme-ui'
 
-import { ListDefinition } from '../../components/ListDefinition'
+import { DefinitionList } from '../../components/DefinitionList'
 import { interpolate } from '../../helpers/interpolate'
 import { splitEvents, VaultHistoryEvent } from './vaultHistory'
 import { getHistoryEventTranslation, VaultHistoryEntry } from './VaultHistoryEntry'
@@ -322,7 +322,7 @@ export function VaultHistoryView({ vaultHistory }: { vaultHistory: VaultHistoryE
           <Heading variant="headerSettings" sx={{ mb: 3 }}>
             {t('vault-history')}
           </Heading>
-          <ListDefinition>
+          <DefinitionList>
             {spitedEvents.map((item) => (
               <VaultHistoryEntry
                 item={item}
@@ -331,7 +331,7 @@ export function VaultHistoryView({ vaultHistory }: { vaultHistory: VaultHistoryE
                 key={`${item.id}${`-${item.splitId}` || ''}`}
               />
             ))}
-          </ListDefinition>
+          </DefinitionList>
         </Card>
       )}
     </>
